move tests out of package, fix build bugs, add coveralls
