+++ .github/workflows/cov_doc.yml
@@ -19,13 +19,16 @@ jobs:
     - uses: actions/setup-python@v2
       with:
         python-version: 3.8
+    - run: pre-commit run -a
     - run: |
         python -m pip install --upgrade pip wheel
         python -m pip install -r requirements-dev.txt
         python -m ipykernel install --user --name python3
-    - run: COVERAGE=1 make
-    - run: make flake8
-    - run: make doc
+        python -m pip install -e . -v
+    - run: cd doc; make
     - run: python -m pytest tutorial
-    - run: make cov
+    - run: python -m pytest --cov src/iminuit
     - uses: codecov/codecov-action@v1
+    - uses: AndreMiras/coveralls-python-action@develop
+      with:
+        github-token: ${{ secrets.GITHUB_TOKEN }}
