+++ setup.cfg
@@ -1,5 +1,5 @@
 [tool:pytest]
-testpaths = src/iminuit
+testpaths = tests
 
 [aliases]
 test=pytest
