+++ tests/test_color.py
