+++ tests/test_cost.py
