+++ tests/test_deprecated.py
